move test classes to dedicated module
