+++ test/test_fields.py
@@ -3,44 +3,13 @@ module for testing functionality of serializable fields
 """
 
 # src
-from objectfactory import Factory, Serializable, Field, Nested, List
-from .test_serializable import MyBasicClass
-
-
-@Factory.register_class
-class MyBasicClassWithLists( Serializable ):
-    """
-    basic class to be used for testing serialization of primitive lists
-    """
-    str_list_prop = Field()
-    int_list_prop = Field()
-
-
-@Factory.register_class
-class MyComplexClass( Serializable ):
-    """
-    complex class to test hierarchical serialization
-    """
-    nested = Nested( MyBasicClass )
-    prop = Field()
-
-
-@Factory.register_class
-class MyOtherComplexClass( Serializable ):
-    """
-    complex class to test list serialization
-    """
-    str_prop = Field()
-    nested_list_prop = List()
-
-
-@Factory.register_class
-class MyClassWithFieldOptionals( Serializable ):
-    """
-    complex class to test list serialization
-    """
-    str_prop = Field( default='default_val' )
-    int_prop = Field( name='int_prop_named' )
+from .testmodule.testclasses import (
+    MyBasicClass,
+    MyComplexClass,
+    MyOtherComplexClass,
+    MyBasicClassWithLists,
+    MyClassWithFieldOptionals
+)
 
 
 class TestNested( object ):
